Save elapsed time

+++ app/src/main/java/com/example/stopwatch/view/binding/BindingAdapter.kt
@@ -5,7 +5,7 @@ import androidx.databinding.BindingAdapter
 import com.example.stopwatch.R
 
 @BindingAdapter("timeElapsed")
-fun TextView.timeElapsed(timeElapsed: Int) {
+fun TextView.timeElapsed(timeElapsed: Long) {
     val hours = timeElapsed / 3600
     val minutes = (timeElapsed - (hours * 3600)) / 60
     val seconds = timeElapsed - (hours * 3600) - (minutes * 60)
